+++ main/webserver_lock.cpp
@@ -1,0 +1,3 @@
+#include "webserver_lock.h"
+
+cpputils::DelayedConstruction<espcpputils::mutex_semaphore> webserver_lock;
